# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/projects:
    get:
      operationId: listProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: List all registered projects (active + degraded)
      tags:
      - projects
    post:
      operationId: addProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Register a new project from a git repository path
      tags:
      - projects
  /api/v1/projects/{id}:
    delete:
      operationId: removeProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveProjectResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Remove a project; stops sessions, cleans workspaces, unregisters
      tags:
      - projects
    get:
      operationId: getProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGetResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Fetch one project; discriminates ok vs degraded
      tags:
      - projects
    put:
      operationId: updateProjectSettings
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectSettingsInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Atomically replace a project's display name and config
      tags:
      - projects
  /api/v1/projects/{id}/config:
    put:
      operationId: setProjectConfig
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetProjectConfigInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Replace a project's per-project config
      tags:
      - projects
  /api/v1/projects/initialize:
    post:
      operationId: initializeProjectRepository
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitializeRepositoryInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitializeRepositoryResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Initialize a selected folder as a Git repository with an initial commit
      tags:
      - projects
...
  /api/v1/remote-hosts:
    get:
      operationId: listRemoteHosts
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRemoteHostsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List registered remote daemon hosts and their connection health
      tags:
      - remoteHosts
    post:
      operationId: registerRemoteHost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRemoteHostInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteHostResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Register a reachable remote daemon host
      tags:
      - remoteHosts
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
    AddProjectInput:
      properties:
        asWorkspace:
          type: boolean
        config:
          $ref: '#/components/schemas/ProjectConfig'
        name:
          type:
          - "null"
          - string
        path:
          type: string
        projectId:
          type:
          - "null"
          - string
      required:
      - path
      type: object
    AgentConfig:
      properties:
        mode:
          type: string
        model:
          type: string
        permissions:
          type: string
      type: object
...
    ContainerReapConfig:
      properties:
        disabled:
          type: boolean
      type: object
...
    DegradedProject:
      properties:
        id:
          type: string
        kind:
          enum:
          - single_repo
          - workspace
          - scratch
          type: string
        name:
          type: string
        path:
          type: string
        resolveError:
          type: string
      required:
      - id
      - name
      - kind
      - path
      - resolveError
      type: object
...
    DomainReviewerConfig:
      properties:
        harness:
          type: string
      required:
      - harness
      type: object
...
    InitializeRepositoryInput:
      properties:
        path:
          type: string
      required:
      - path
      type: object
    InitializeRepositoryResult:
      properties:
        path:
          type: string
      required:
      - path
      type: object
...
    ListProjectsResponse:
      properties:
        projects:
          items:
            $ref: '#/components/schemas/ProjectSummary'
          type: array
        remoteHostInventoryError:
          type: string
        remoteHostInventoryStale:
          type: boolean
        remoteHosts:
          items:
            $ref: '#/components/schemas/RemoteHost'
          type: array
      required:
      - projects
      type: object
    ListRemoteHostsResponse:
      properties:
        remoteHosts:
          items:
            $ref: '#/components/schemas/RemoteHost'
          type: array
      required:
      - remoteHosts
      type: object
...
    Project:
      properties:
        agent:
          type: string
        config:
          $ref: '#/components/schemas/ProjectConfig'
        defaultBranch:
          type: string
        id:
          type: string
        kind:
          enum:
          - single_repo
          - workspace
          - scratch
          type: string
        name:
          type: string
        path:
          type: string
        repo:
          type: string
        workspaceRepos:
          items:
            $ref: '#/components/schemas/WorkspaceRepo'
          type: array
      required:
      - id
      - name
      - kind
      - path
      - repo
      - defaultBranch
      type: object
    ProjectConfig:
      properties:
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
        agentRules:
          type: string
        agentRulesFile:
          type: string
        containerReap:
          $ref: '#/components/schemas/ContainerReapConfig'
        defaultBranch:
          type: string
        env:
          additionalProperties:
            type: string
          type: object
        orchestrator:
          $ref: '#/components/schemas/RoleOverride'
        orchestratorRules:
          type: string
        postCreate:
          items:
            type: string
          type: array
        reviewers:
          items:
            $ref: '#/components/schemas/DomainReviewerConfig'
          type: array
        sessionPrefix:
          type: string
        symlinks:
          items:
            type: string
          type: array
        trackerIntake:
          $ref: '#/components/schemas/TrackerIntakeConfig'
        worker:
          $ref: '#/components/schemas/RoleOverride'
      type: object
    ProjectGetResponse:
      properties:
        project:
          $ref: '#/components/schemas/ProjectOrDegraded'
        status:
          enum:
          - ok
          - degraded
          type: string
      required:
      - status
      - project
      type: object
    ProjectOrDegraded:
      oneOf:
      - $ref: '#/components/schemas/Project'
      - $ref: '#/components/schemas/DegradedProject'
      type: object
    ProjectResponse:
      properties:
        project:
          $ref: '#/components/schemas/Project'
      required:
      - project
      type: object
    ProjectSource:
      properties:
        available:
          type: boolean
        hostId:
          type: string
        kind:
          type: string
        name:
          type: string
        path:
          type: string
        unavailableReason:
          type: string
      required:
      - available
      - name
      - path
      - kind
      type: object
    ProjectSummary:
      properties:
        id:
          type: string
        kind:
          enum:
          - single_repo
          - workspace
          - scratch
          type: string
        metadataConflicts:
          items:
            type: string
          type: array
        name:
          type: string
        orchestratorAgent:
          type: string
        path:
          type: string
        resolveError:
          type: string
        sessionPrefix:
          type: string
        sources:
          items:
            $ref: '#/components/schemas/ProjectSource'
          type: array
      required:
      - id
      - name
      - path
      - kind
      - sessionPrefix
      type: object
...
    RegisterRemoteHostInput:
      properties:
        address:
          type: string
        hostId:
          type: string
        label:
          type: string
      required:
      - hostId
      - address
      type: object
...
    RemoteHost:
      properties:
        address:
          type: string
        createdAt:
          format: date-time
          type: string
        hostId:
          type: string
        inventoryError:
          type: string
        inventoryStale:
          type: boolean
        label:
          type: string
        lastProbeAt:
          format: date-time
          type: string
        lastProbeError:
          type: string
        lastProbeSucceeded:
          type: boolean
        state:
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - hostId
      - address
      - state
      - lastProbeAt
      - lastProbeSucceeded
      - createdAt
      - updatedAt
      type: object
    RemoteHostResponse:
      properties:
        remoteHost:
          $ref: '#/components/schemas/RemoteHost'
      required:
      - remoteHost
      type: object
...
    RemoveProjectResult:
      properties:
        projectId:
          type: string
        removedStorageDir:
          type: boolean
      required:
      - projectId
      - removedStorageDir
      type: object
...
    RoleOverride:
      properties:
        agent:
          type: string
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
      type: object
...
    SetProjectConfigInput:
      properties:
        config:
          $ref: '#/components/schemas/ProjectConfig'
      required:
      - config
      type: object
...
    TrackerIntakeConfig:
      properties:
        assignee:
          type: string
        enabled:
          type: boolean
        provider:
          enum:
          - github
          type: string
        repo:
          type: string
      type: object
...
    UpdateProjectSettingsInput:
      properties:
        config:
          $ref: '#/components/schemas/ProjectConfig'
        displayName:
          maxLength: 20
          minLength: 1
          type: string
      required:
      - displayName
      - config
      type: object
...
    WorkspaceRepo:
      properties:
        name:
          type: string
        relativePath:
          type: string
        repo:
          type: string
      required:
      - name
      - relativePath
      - repo
      type: object
...